+++ docs/jangar/architecture-inventory.md
@@ -12,28 +12,28 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 
 ## Top 20 Largest App Modules
 
-| Rank | Module                                                                 |  LOC |
-| ---- | ---------------------------------------------------------------------- | ---: |
-| 1    | `services/jangar/src/data/agents-control-plane.ts`                     | 3382 |
-| 2    | `services/jangar/src/server/supporting-primitives-controller.ts`       | 3346 |
-| 3    | `services/jangar/src/server/codex-judge.ts`                            | 2729 |
-| 4    | `services/jangar/src/server/atlas-store.ts`                            | 2148 |
-| 5    | `services/jangar/src/server/orchestration-controller.ts`               | 2138 |
-| 6    | `services/jangar/src/server/chat-completion-encoder.ts`                | 2076 |
-| 7    | `services/jangar/src/server/torghut-market-context-agents.ts`          | 1981 |
-| 8    | `services/jangar/src/server/agents-controller/index.ts`                | 1828 |
-| 9    | `services/jangar/src/server/torghut-simulation-control-plane.ts`       | 1741 |
-| 10   | `services/jangar/src/server/github-review-store.ts`                    | 1459 |
-| 11   | `services/jangar/src/server/codex-judge-store.ts`                      | 1329 |
-| 12   | `services/jangar/src/server/torghut-whitepapers.ts`                    | 1328 |
-| 13   | `services/jangar/src/server/chat.ts`                                   | 1313 |
-| 14   | `services/jangar/src/server/agents-controller/agent-run-reconciler.ts` | 1191 |
-| 15   | `services/jangar/src/server/agentctl-grpc.ts`                          | 1171 |
-| 16   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx`     | 1028 |
-| 17   | `services/jangar/src/server/torghut-trading.ts`                        |  988 |
-| 18   | `services/jangar/src/server/db.ts`                                     |  974 |
-| 19   | `services/jangar/src/routes/torghut/trading.tsx`                       |  952 |
-| 20   | `services/jangar/src/server/torghut-quant-metrics.ts`                  |  929 |
+| Rank | Module                                                             |  LOC |
+| ---- | ------------------------------------------------------------------ | ---: |
+| 1    | `services/jangar/src/data/agents-control-plane.ts`                 | 3382 |
+| 2    | `services/jangar/src/server/supporting-primitives-controller.ts`   | 3346 |
+| 3    | `services/jangar/src/server/codex-judge.ts`                        | 2729 |
+| 4    | `services/jangar/src/server/atlas-store.ts`                        | 2148 |
+| 5    | `services/jangar/src/server/orchestration-controller.ts`           | 2138 |
+| 6    | `services/jangar/src/server/chat-completion-encoder.ts`            | 2076 |
+| 7    | `services/jangar/src/server/torghut-market-context-agents.ts`      | 1981 |
+| 8    | `services/jangar/src/server/agents-controller/index.ts`            | 1824 |
+| 9    | `services/jangar/src/server/torghut-simulation-control-plane.ts`   | 1741 |
+| 10   | `services/jangar/src/server/github-review-store.ts`                | 1459 |
+| 11   | `services/jangar/src/server/codex-judge-store.ts`                  | 1329 |
+| 12   | `services/jangar/src/server/torghut-whitepapers.ts`                | 1328 |
+| 13   | `services/jangar/src/server/chat.ts`                               | 1313 |
+| 14   | `services/jangar/src/server/agentctl-grpc.ts`                      | 1171 |
+| 15   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx` | 1028 |
+| 16   | `services/jangar/src/server/torghut-trading.ts`                    |  988 |
+| 17   | `services/jangar/src/server/db.ts`                                 |  974 |
+| 18   | `services/jangar/src/routes/torghut/trading.tsx`                   |  952 |
+| 19   | `services/jangar/src/server/torghut-quant-metrics.ts`              |  929 |
+| 20   | `services/jangar/src/server/github-review-actions.ts`              |  928 |
 
 ## Control-Plane Route Surface
